+++ packages/zosmf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OSMF SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added `tsoAccount` and `tsoProcedure` options to the zosmf profile/schema, allowing users to specify a TSO account number and logon procedure to use for z/OSMF data set and file operations. [#2653](https://github.com/zowe/zowe-cli/issues/2653)
+
 ## `8.32.0`
 
 - Enhancement: Added `certAccount` option to z/OSMF connection options to support client certificate authentication using certificates stored in system keystores on macOS and Windows platforms.